+++ src/bask_project.h
@@ -3,7 +3,7 @@
 
 struct bask_project
 {
-	int p_active;
+	short p_active;
 	float p_complete;
 	float p_tasks;
 	char p_name[T_S_PROJECT];
@@ -11,6 +11,6 @@ struct bask_project
 };
 
 void project_free_ll (struct bask_project**);
-int project_insert (struct bask_project**, int, int, char*);
+int project_insert (struct bask_project**, short, int, char*);
 
 #endif /* __BASK_PROJECT_H__ */
